Refactor page details into a separate module
Fixes #509.
