+++ src/.eslintrc.json
@@ -3,7 +3,7 @@
     "react"
   ],
   "parserOptions": {
-    "ecmaVersion": 8,
+    "ecmaVersion": 2020,
     "ecmaFeatures": {
       "jsx": true
     }
