+++ src/components/page-url-details/page-url-details.css
@@ -1,0 +1,31 @@
+.page-url-details {
+  padding: 0 1em;
+}
+
+.page-url-details summary {
+  cursor: pointer;
+  margin-left: -1em;
+}
+
+.page-url-details ins {
+  font-weight: bold;
+  background: none !important;
+}
+
+.page-url-details a {
+  /*
+    Links here are full URLs are can sometimes have no good line break points.
+  */
+  line-break: anywhere;
+}
+
+.url-history-list {
+  list-style-type: none;
+  margin-left: 0;
+  padding-left: 0;
+}
+
+.url-scheme {
+  display: inline-block;
+  width: 6em;
+}
